requirements.txt
accelerate==1.8.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
build==1.2.2.post1
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chromadb==1.0.13
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
effdet==0.4.1
email_validator==2.2.0
emoji==2.14.1
exceptiongroup==1.3.0
fastapi==0.115.13
fastapi-cli==0.0.7
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
hf-xet==1.1.4
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kombu==5.5.4
kubernetes==33.1.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.4.8
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.5.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
openai==1.88.0
opencv-python==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==0.22.0
pikepdf==9.9.0
pillow==11.2.1
posthog==5.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pypdf==5.6.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.36.0
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
